m365_wrapper/requirements.txt
microsoft-agents-hosting-core>=0.8.0
microsoft-agents-hosting-fastapi>=0.8.0
microsoft-agents-authentication-msal>=0.8.0
msal>=1.31.1
PyJWT[crypto]>=2.10.1
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
aiohttp>=3.9.0
